Move Operator and AbstractOperator classes to cluster-operator module
