+++ cluster-operator/pom.xml
@@ -58,6 +58,10 @@
         </dependency>
         <dependency>
             <groupId>io.fabric8</groupId>
+            <artifactId>kubernetes-model-common</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.fabric8</groupId>
             <artifactId>kubernetes-model-events</artifactId>
         </dependency>
         <dependency>
